Discord: Simplify and refactor bot
